fix(cond): Add lastHeartbeatTime to conditions
lastHeartbeatTime is meant to track the last time that an update was
given to a particular condition. One scenario where this is useful is
where an operator is managing a resource that is in a "Degraded" state
lastHeartbeatTime would let an admin know how recently that state has
been observed.

Unit tests were updated to test the change in logic.

Fixes #3