SetStatusCondition[NoHeartbeat]: return 'changed' bool
Add a boolean return value to SetStatusCondition and
SetStatusConditionNoHeartbeat indicating if the function made any
changes to the conditions (other than the heartbeat timestamp in the
case of the former). This allows consumers to avoid thrashing by
requeueing an object that has not changed substantively.

This represents an additive API change which should be backward
compatible as existing consumers will simply be ignoring the new return
value.